% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathintegrations.R
\name{trapezoidal_rule}
\alias{trapezoidal_rule}
\title{Math integrations}
\usage{
trapezoidal_rule(original_x, original_y)
}
\description{
Extension of the fda.usc R package for mathematical integrations/interpolations
}
