% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterstorageprofile.R
\name{waterstorageprofile}
\alias{waterstorageprofile}
\title{Water storage}
\usage{
waterstorageprofile()
}
\description{
This function is used to calculate the water storage index of a given point in soil
Using three different mathematical methods.
}
